Refacto: rename addProductForm - addProductSection, normalize admin section patern
